Move /op texts to texts.js, and add the thick force
This commit replaces the inlined text for /op into texts.js. Text for /own is also added.

Additionally, the thick force is added, which prevents you from using negative commands on people with a higher op level than you. Commands can set the verb displayed by the thick force, which defaults to "verbing". A percentage sign in the verb is replaced with the target, otherwise it is appended after a space.